FEAT: Added a new class to customize Page Ports and added 2 new properties  (#6374)
