+++ SECURITY.md
@@ -22,7 +22,9 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 -->
 
- ## Reporting a vulnerability
+# Security Policy
+
+## Reporting a vulnerability
 
 > [!CAUTION]
 > Do not use GitHub issues to report any security vulnerabilities.
@@ -32,7 +34,7 @@ mentioning the repository and the details of your finding. The team will address
 
 Provide the PyAnsys Core team with this information:
 
- - Any specific configuration settings needed to reproduce the problem
- - Step-by-step guidance to reproduce the problem
- - The exact location of the problematic source code, including tag, branch, commit, or a direct URL
- - The potential consequences of the vulnerability, along with a description of how an attacker could take advantage of the issue
+- Any specific configuration settings needed to reproduce the problem
+- Step-by-step guidance to reproduce the problem
+- The exact location of the problematic source code, including tag, branch, commit, or a direct URL
+- The potential consequences of the vulnerability, along with a description of how an attacker could take advantage of the issue
