+++ doc/changelog.d/6374.added.md
@@ -1,0 +1,1 @@
+Added a new class to customize page ports and added 2 new properties
