+++ src/ansys/aedt/core/modules/boundary/circuit_boundary.py
@@ -1282,413 +1282,3 @@ class CurrentSinSource(Sources, object):
     def tone(self, value):
         self._props["TONE"] = value
         self._child.SetPropValue("TONE", value)
-
-
-class Excitations(object):
-    """Manages Excitations in Circuit Projects.
-
-    Examples
-    --------
-
-    """
-
-    def __init__(self, app, name):
-        self._app = app
-        self._name = name
-        for comp in self._app.modeler.schematic.components:
-            if (
-                "PortName" in self._app.modeler.schematic.components[comp].parameters.keys()
-                and self._app.modeler.schematic.components[comp].parameters["PortName"] == self.name
-            ):
-                self.schematic_id = comp
-                self.id = self._app.modeler.schematic.components[comp].id
-                self._angle = self._app.modeler.schematic.components[comp].angle
-                self.levels = self._app.modeler.schematic.components[comp].levels
-                self._location = self._app.modeler.schematic.components[comp].location
-                self._mirror = self._app.modeler.schematic.components[comp].mirror
-                self.pins = self._app.modeler.schematic.components[comp].pins
-                self._use_symbol_color = self._app.modeler.schematic.components[comp].usesymbolcolor
-                break
-        self._props = self._excitation_props(name)
-        self._auto_update = True
-
-    @property
-    def name(self):
-        """Excitation name.
-
-        Returns
-        -------
-        str
-        """
-        return self._name
-
-    @name.setter
-    def name(self, port_name):
-        if port_name not in self._app.excitation_names:
-            if port_name != self._name:
-                # Take previous properties
-                self._app.odesign.RenamePort(self._name, port_name)
-                self._name = port_name
-                self._app.modeler.schematic.components[self.schematic_id].name = "IPort@" + port_name
-                self.pins[0].name = "IPort@" + port_name + ";" + str(self.schematic_id)
-        else:
-            self._logger.warning("Name %s already assigned in the design", port_name)
-
-    @property
-    def angle(self):
-        """Symbol angle.
-
-        Returns
-        -------
-        float
-        """
-        return self._angle
-
-    @angle.setter
-    def angle(self, angle):
-        self._app.modeler.schematic.components[self.schematic_id].angle = angle
-
-    @property
-    def mirror(self):
-        """Enable port mirror.
-
-        Returns
-        -------
-        bool
-        """
-        return self._mirror
-
-    @mirror.setter
-    def mirror(self, mirror_value=True):
-        self._app.modeler.schematic.components[self.schematic_id].mirror = mirror_value
-        self._mirror = mirror_value
-
-    @property
-    def location(self):
-        """Port location.
-
-        Returns
-        -------
-        list
-        """
-        return self._location
-
-    @location.setter
-    def location(self, location_xy):
-        # The command must be called two times.
-        self._app.modeler.schematic.components[self.schematic_id].location = location_xy
-        self._app.modeler.schematic.components[self.schematic_id].location = location_xy
-        self._location = location_xy
-
-    @property
-    def use_symbol_color(self):
-        """Use symbol color.
-
-        Returns
-        -------
-        list
-        """
-        return self._use_symbol_color
-
-    @use_symbol_color.setter
-    def use_symbol_color(self, use_color=True):
-        self._app.modeler.schematic.components[self.schematic_id].usesymbolcolor = use_color
-        self._app.modeler.schematic.components[self.schematic_id].set_use_symbol_color(use_color)
-        self._use_symbol_color = use_color
-
-    @property
-    def impedance(self):
-        """Port termination.
-
-        Returns
-        -------
-        list
-        """
-        return [self._props["rz"], self._props["iz"]]
-
-    @impedance.setter
-    def impedance(self, termination=None):
-        if termination and len(termination) == 2:
-            self._app.modeler.schematic.components[self.schematic_id].change_property(
-                ["NAME:rz", "Value:=", termination[0]]
-            )
-            self._app.modeler.schematic.components[self.schematic_id].change_property(
-                ["NAME:iz", "Value:=", termination[1]]
-            )
-            self._props["rz"] = termination[0]
-            self._props["iz"] = termination[1]
-
-    @property
-    def enable_noise(self):
-        """Enable noise.
-
-        Returns
-        -------
-        bool
-        """
-
-        return self._props["EnableNoise"]
-
-    @enable_noise.setter
-    def enable_noise(self, enable=False):
-        self._app.modeler.schematic.components[self.schematic_id].change_property(
-            ["NAME:EnableNoise", "Value:=", enable]
-        )
-        self._props["EnableNoise"] = enable
-
-    @property
-    def noise_temperature(self):
-        """Enable noise.
-
-        Returns
-        -------
-        str
-        """
-
-        return self._props["noisetemp"]
-
-    @noise_temperature.setter
-    def noise_temperature(self, noise=None):
-        if noise:
-            self._app.modeler.schematic.components[self.schematic_id].change_property(
-                ["NAME:noisetemp", "Value:=", noise]
-            )
-            self._props["noisetemp"] = noise
-
-    @property
-    def microwave_symbol(self):
-        """Enable microwave symbol.
-
-        Returns
-        -------
-        bool
-        """
-        if self._props["SymbolType"] == 1:
-            return True
-        else:
-            return False
-
-    @microwave_symbol.setter
-    def microwave_symbol(self, enable=False):
-        if enable:
-            self._props["SymbolType"] = 1
-        else:
-            self._props["SymbolType"] = 0
-        self.update()
-
-    @property
-    def reference_node(self):
-        """Reference node.
-
-        Returns
-        -------
-        str
-        """
-        if self._props["RefNode"] == "Z":
-            return "Ground"
-        return self._props["RefNode"]
-
-    @reference_node.setter
-    def reference_node(self, ref_node=None):
-        if ref_node:
-            self._logger.warning("Set reference node only working with gRPC")
-            if ref_node == "Ground":
-                ref_node = "Z"
-            self._props["RefNode"] = ref_node
-            self.update()
-
-    @property
-    def enabled_sources(self):
-        """Enabled sources.
-
-        Returns
-        -------
-        list
-        """
-        return self._props["EnabledPorts"]
-
-    @enabled_sources.setter
-    def enabled_sources(self, sources=None):
-        if sources:
-            self._props["EnabledPorts"] = sources
-            self.update()
-
-    @property
-    def enabled_analyses(self):
-        """Enabled analyses.
-
-        Returns
-        -------
-        dict
-        """
-        return self._props["EnabledAnalyses"]
-
-    @enabled_analyses.setter
-    def enabled_analyses(self, analyses=None):
-        if analyses:
-            self._props["EnabledAnalyses"] = analyses
-            self.update()
-
-    @pyaedt_function_handler()
-    def _excitation_props(self, port):
-        excitation_prop_dict = {}
-        for comp in self._app.modeler.schematic.components:
-            if (
-                "PortName" in self._app.modeler.schematic.components[comp].parameters.keys()
-                and self._app.modeler.schematic.components[comp].parameters["PortName"] == port
-            ):
-                excitation_prop_dict["rz"] = "50ohm"
-                excitation_prop_dict["iz"] = "0ohm"
-                excitation_prop_dict["term"] = None
-                excitation_prop_dict["TerminationData"] = None
-                excitation_prop_dict["RefNode"] = "Z"
-                excitation_prop_dict["EnableNoise"] = False
-                excitation_prop_dict["noisetemp"] = "16.85cel"
-
-                if "RefNode" in self._app.modeler.schematic.components[comp].parameters:
-                    excitation_prop_dict["RefNode"] = self._app.modeler.schematic.components[comp].parameters["RefNode"]
-                if "term" in self._app.modeler.schematic.components[comp].parameters:
-                    excitation_prop_dict["term"] = self._app.modeler.schematic.components[comp].parameters["term"]
-                    excitation_prop_dict["TerminationData"] = self._app.modeler.schematic.components[comp].parameters[
-                        "TerminationData"
-                    ]
-                else:
-                    if "rz" in self._app.modeler.schematic.components[comp].parameters:
-                        excitation_prop_dict["rz"] = self._app.modeler.schematic.components[comp].parameters["rz"]
-                        excitation_prop_dict["iz"] = self._app.modeler.schematic.components[comp].parameters["iz"]
-
-                if "EnableNoise" in self._app.modeler.schematic.components[comp].parameters:
-                    if self._app.modeler.schematic.components[comp].parameters["EnableNoise"] == "true":
-                        excitation_prop_dict["EnableNoise"] = True
-                    else:
-                        excitation_prop_dict["EnableNoise"] = False
-
-                    excitation_prop_dict["noisetemp"] = self._app.modeler.schematic.components[comp].parameters[
-                        "noisetemp"
-                    ]
-
-                if not self._app.design_properties or not self._app.design_properties["NexximPorts"]["Data"]:
-                    excitation_prop_dict["SymbolType"] = 0
-                else:
-                    excitation_prop_dict["SymbolType"] = self._app.design_properties["NexximPorts"]["Data"][port][
-                        "SymbolType"
-                    ]
-
-                if "pnum" in self._app.modeler.schematic.components[comp].parameters:
-                    excitation_prop_dict["pnum"] = self._app.modeler.schematic.components[comp].parameters["pnum"]
-                else:
-                    excitation_prop_dict["pnum"] = None
-                source_port = []
-                if not self._app.design_properties:
-                    enabled_ports = None
-                else:
-                    enabled_ports = self._app.design_properties["ComponentConfigurationData"]["EnabledPorts"]
-                if isinstance(enabled_ports, dict):
-                    for source in enabled_ports:
-                        if enabled_ports[source] and port in enabled_ports[source]:
-                            source_port.append(source)
-                excitation_prop_dict["EnabledPorts"] = source_port
-
-                components_port = []
-                if not self._app.design_properties:
-                    multiple = None
-                else:
-                    multiple = self._app.design_properties["ComponentConfigurationData"]["EnabledMultipleComponents"]
-                if isinstance(multiple, dict):
-                    for source in multiple:
-                        if multiple[source] and port in multiple[source]:
-                            components_port.append(source)
-                excitation_prop_dict["EnabledMultipleComponents"] = components_port
-
-                port_analyses = {}
-                if not self._app.design_properties:
-                    enabled_analyses = None
-                else:
-                    enabled_analyses = self._app.design_properties["ComponentConfigurationData"]["EnabledAnalyses"]
-                if isinstance(enabled_analyses, dict):
-                    for source in enabled_analyses:
-                        if (
-                            enabled_analyses[source]
-                            and port in enabled_analyses[source]
-                            and source in excitation_prop_dict["EnabledPorts"]
-                        ):
-                            port_analyses[source] = enabled_analyses[source][port]
-                excitation_prop_dict["EnabledAnalyses"] = port_analyses
-                return excitation_prop_dict
-
-    @pyaedt_function_handler()
-    def update(self):
-        """Update the excitation in AEDT.
-
-        Returns
-        -------
-        bool
-            ``True`` when successful, ``False`` when failed.
-
-        """
-
-        # self._logger.warning("Property port update only working with GRPC")
-
-        if self._props["RefNode"] == "Ground":
-            self._props["RefNode"] = "Z"
-
-        arg0 = [
-            "NAME:" + self.name,
-            "IIPortName:=",
-            self.name,
-            "SymbolType:=",
-            self._props["SymbolType"],
-            "DoPostProcess:=",
-            False,
-        ]
-
-        arg1 = ["NAME:ChangedProps"]
-        arg2 = []
-
-        # Modify RefNode
-        if self._props["RefNode"] != "Z":
-            arg2 = [
-                "NAME:NewProps",
-                ["NAME:RefNode", "PropType:=", "TextProp", "OverridingDef:=", True, "Value:=", self._props["RefNode"]],
-            ]
-
-        # Modify Termination
-        if self._props["term"] and self._props["TerminationData"]:
-            arg2 = [
-                "NAME:NewProps",
-                ["NAME:term", "PropType:=", "TextProp", "OverridingDef:=", True, "Value:=", self._props["term"]],
-            ]
-
-        for prop in self._props:
-            skip1 = (prop == "rz" or prop == "iz") and isinstance(self._props["term"], str)
-            skip2 = prop == "EnabledPorts" or prop == "EnabledMultipleComponents" or prop == "EnabledAnalyses"
-            skip3 = prop == "SymbolType"
-            skip4 = prop == "TerminationData" and not isinstance(self._props["term"], str)
-            if not skip1 and not skip2 and not skip3 and not skip4 and self._props[prop] is not None:
-                command = ["NAME:" + prop, "Value:=", self._props[prop]]
-                arg1.append(command)
-
-        arg1 = [["NAME:Properties", arg2, arg1]]
-        self._app.odesign.ChangePortProperty(self.name, arg0, arg1)
-
-        for source in self._app.sources:
-            self._app.sources[source].update()
-        return True
-
-    @pyaedt_function_handler()
-    def delete(self):
-        """Delete the port in AEDT.
-
-        Returns
-        -------
-        bool
-            ``True`` when successful, ``False`` when failed.
-
-        """
-        self._app.modeler._odesign.DeletePort(self.name)
-        return True
-
-    @property
-    def _logger(self):
-        """Logger."""
-        return self._app.logger
